@@ -4,9 +4,9 @@ class Downloader
   end
 
   def download(file_name)
-    s3 = AWS::S3.new
-    bucket = s3.buckets[@bucket]
-    object = bucket.objects[file_name]
+    s3 = Aws::S3::Resource.new
+    bucket = s3.bucket(@bucket)
+    object = bucket.object(file_name)
 
     buffer = StringIO.new("", 'w')
     object.read do |chunk|
@@ -18,9 +18,9 @@ class Downloader
   end
 
   def delete_file(name)
-    s3 = AWS::S3.new
-    bucket = s3.buckets[@bucket]
-    object = bucket.objects[name]
+    s3 = Aws::S3::Resource.new
+    bucket = s3.bucket(@bucket)
+    object = bucket.object(name)
     object.delete
 
     return !object.exists?
