@@ -2,18 +2,18 @@ class Receiver
   attr_reader :sqs, :count
 
   def initialize(queue_name)
-    @sqs = AWS::SQS.new
+    @sqs = Aws::SQS::Resource.new
     @limit = 10
     @queue_name = queue_name
     @count = 0
   end
 
   def receive_messages
-    queue = sqs.queues.named(@queue_name)
-
+    queue = @sqs.queue(@queue_name)
 
     4.times do
-      queue.receive_message(limit: @limit) do |sqs_message|
+      messages = queue.receive_messages(max_number_of_messages: @limit)
+      messages.each do |sqs_message|
         msg = Messages::MessageParser.parse(sqs_message)
 
         next if msg.empty?
