@@ -2,13 +2,13 @@ class Sender
   attr_reader :sqs
 
   def initialize(bucket)
-    @sqs = AWS::SQS.new
+    @sqs = Aws::SQS::Resource.new
     @queue_name = bucket
   end
 
   def send_message(event_message)
-    queue = sqs.queues.named(@queue_name)
-    sqs_message = queue.send_message(event_message.to_xml)
-    sqs_message.id
+    queue = sqs.queue(@queue_name)
+    sqs_message = queue.send_message({message_body: event_message.to_xml})
+    sqs_message.message_id
   end
 end
